pyproject.toml
[project]
name = "pi-robot"
version = "0.1.0"
description = ""
authors = [
    {name = "jimming"}
]
readme = "README.md"
package-mode = false
requires-python = "^3.10"
dependencies = [
    "openai[realtime] (>=1.63.0,<2.0.0)",
    "pyaudio (>=0.2.14,<0.3.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "gpiozero (>=2.0.1,<3.0.0)",
    "scipy (>=1.15.2,<2.0.0)",
    "rpi-gpio (>=0.7.1,<0.8.0)",
    "lgpio (>=0.2.2.0,<0.3.0.0)",
    "scooterbot-agent (>=1.0.5,<2.0.0)",
    "vosk (>=0.3.45,<0.4.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
flake8 = "^7.1.1"
types-pyaudio = "^0.2.16.20240516"
scipy-stubs = "^1.15.2.0"
types-pyyaml = "^6.0.12.20241230"

